Logic for LRU included in arena logic Some casting added Some grammar corrected

+++ src/lib/vector/vector.h
@@ -9,9 +9,9 @@
 
 #define VECTOR_INIT(vec) vector vec; vector_init(&vec, 0)
 #define VECTOR_ADD(vec, item) vector_add(&vec, (void *) item)
-#define VECTOR_SET(vec, id, item) vector_set(&vec, id, (void *) item)
-#define VECTOR_GET(vec, type, id) (type) vector_get(&vec, id)
-#define VECTOR_DELETE(vec, id) vector_delete(&vec, id)
+#define VECTOR_SET(vec, id, item) vector_set(&vec, (int) id, (void *) item)
+#define VECTOR_GET(vec, type, id) (type) vector_get(&vec, (int) id)
+#define VECTOR_DELETE(vec, id) vector_delete(&vec, (int) id)
 #define VECTOR_TOTAL(vec) vector_total(&vec)
 #define VECTOR_FREE(vec) vector_free(&vec)
